add, move and rename packages to match the java convention
