+++ src/java/by/epam/ultraair/presentation/dto/FlightDTO.java
@@ -1,0 +1,11 @@
+package by.epam.ultraair.presentation.dto;
+
+import java.util.Date;
+
+public class FlightDTO {
+    public String from;
+    public String to;
+    public Date departureDate;
+    public Date arrivalDate;
+    public int passengersAmount;
+}
