+++ src/java/by/epam/ultraair/service/Constants.java
@@ -1,0 +1,7 @@
+package by.epam.ultraair.service;
+
+public interface Constants {
+    double LUGGAGE_TRANSFER_COST = 15.49;
+    int MAX_WEIGHT = 32;
+    int MAX_PASSENGERS = 68;
+}
