--- src/java/presentation/DTO/FlightDTO.java
@@ -1,11 +1,0 @@
-package presentation.dto;
-
-import java.util.Date;
-
-public class FlightDTO {
-    public String from;
-    public String to;
-    public Date departureDate;
-    public Date arrivalDate;
-    public int passengersAmount;
-}
