--- src/java/service/Constants.java
@@ -1,7 +1,0 @@
-package service;
-
-public interface Constants {
-    double LUGGAGE_TRANSFER_COST = 15.49;
-    int MAX_WEIGHT = 32;
-    int MAX_PASSENGERS = 68;
-}
